rd-74 support html and jsx
