+++ packages/robindoc/src/types/content.ts
@@ -5,7 +5,8 @@ export type RobinProps = {
 };
 
 export type Components = {
-    [key: string]: (props: Record<string, string | true | undefined | React.ReactNode>) => JSX.Element;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    [key: string]: (props: Record<string, any>) => JSX.Element;
 };
 
 export type DocsTree = { origPath: string; clientPath: string }[];
